Moves kotest examples to docs
