+++ docs/examples/kotest/build.gradle
@@ -14,9 +14,10 @@ repositories {
 }
 
 dependencies {
-    testImplementation 'org.testcontainers:postgresql'
-    testImplementation("io.kotest:kotest-runner-junit5-jvm:$kotestVersion")
     implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
+    testCompile project(':testcontainers')
+    testCompile project(':kotest')
+    testImplementation("io.kotest:kotest-runner-junit5-jvm:$kotestVersion")
     testImplementation("io.kotest:kotest-assertions-core-jvm:$kotestVersion")
 }
