+++ docs/examples/kotest/src/test/kotlin/com/example/kotest/TestContainerKotestIntegrationPerTest.kt
@@ -3,13 +3,15 @@ package com.example.kotest
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import kotlinx.coroutines.delay
+import org.testcontainers.containers.ContainerPerTestListener
+import org.testcontainers.containers.GenericContainer
 
-class TestcontainerKotestIntegrationPerSpec : StringSpec({
+class TestContainerKotestIntegrationPerTest : StringSpec({
     val container = GenericContainer<Nothing>("ubuntu")
     setContainerCommand(container) // Set container command for first time to echo hello Kotest
 
     // Register the Listener
-    listeners(ContainerPerSpecListener(container))
+    listeners(ContainerPerTestListener(container))
 
     "test should echo hello Kotest for first test" {
         container.logs.trim() shouldBe "hello Kotest"
@@ -19,8 +21,8 @@ class TestcontainerKotestIntegrationPerSpec : StringSpec({
     }
 
     "test should echo hello Kotest for second spec as well" {
-        // presence of "hello Kotest" verifies that container is not restarted.
-        container.logs.trim() shouldBe "hello Kotest"
+        // presence of hello Kotest again verifies that container is restarted.
+        container.logs.trim() shouldBe "hello Kotest again"
     }
 
 })
